apis: move each API into its own namespaced folder
Sygne and Fregata are now folders/namespaces and now have their own
Api class which deals with API stuff, and can soon incorporate more
resources related to their particular API (most notably mappers).
