+++ app/jobs/fetch_students_job.rb
@@ -16,6 +16,6 @@ class FetchStudentsJob < ApplicationJob
   end
 
   def perform(establishment)
-    StudentApi.fetch_students!(establishment.students_provider, establishment.uai)
+    StudentsApi.fetch_students!(establishment.students_provider, establishment.uai)
   end
 end
