+++ spec/jobs/fetch_student_information_job_spec.rb
@@ -69,7 +69,7 @@ RSpec.describe FetchStudentInformationJob, :student_api do
       before { student.update!(ine_not_found: true) }
 
       it "does not make a request the API" do
-        endpoint = StudentsApi::Sygne.new(establishment.uai).student_endpoint(student.ine)
+        endpoint = StudentsApi::Sygne::Api.new(establishment.uai).student_endpoint(student.ine)
 
         described_class.perform_now(schooling)
